(maint) Gemfile updates
The version of Puppet/Facter/Hiera will determine the exact versions
of the Windows gems, so relax the versioning a bit to allow for the
proper versions of gems to be pulled in.

Explicitly determine when we need to require Windows gems, which is
when we are using a git or file location for Puppet, or are requiring
a Puppet version less than 3.5.0 (the first versions we had Windows
platform specific gems).