(MODULES-2207) Gemfile - Fix explicit windows check
Add fixes for Windows ruby check so that it is  deterministic.
The prior check did not actually keep from requiring windows gems on
non-windows platforms due to the behavior of `platform`, which attaches
itself to the end of each gem in the block. Instead prefer actual
branching logic using a check against `Gem::Platform.local.os`.