(FM-5165) Add MAINTAINERS.md file to all modules
All Puppet projects should contain a Maintenance section as either inside
README or in a separate file called MAINTAINERS.md.  This commit adds the
default modsync to create a MAINTAINERS.md file.  A file was used instead
of updating README as it is much easier to managae and automate the entire
content of the file instead of a section within a README, which is unique per
module.

The default of the Forge Modules mailing address is used, but can be modified
on a per module basis.