Require beaker 3.7 for all modules
Beaker is running on a ruby independent of the SUT, which allows us to
make tighter requirements on the runtime environment. This removes much
complexity from the gem selection.

Now we can require beaker 3.7 across all modules, which has significant
improvements for windows SUTs on vmpooler and vagrant.

This is a cherry-pick of the same commit on the windows side.