(maint) Update Travis CI command
Previously, the Travis CI command used to verify the modsync configurations was
modified to exclue the vsphere and sqlserver modules in commit 301262746cb.
However the command was escaping text which did not need to be, thus the command
was failing to exclude those modules.  This commit removes the escaping and
ensures that the vsphere and sqlserver module are not checked during a Travis CI
run. These modules are excluded as they are private Github repos.