[MODULES-4506] Implement puppet-module-gems
This involves removing a number of gems from the development and system_tests groups in the Gemfile config that have been abstracted out to puppet-module-gems and referencing just the puppet-module-gems themselves.

This is an attempt to reduce the need to run msync every time we have a dependency conflict, by simply updating the puppet-module-gems generated gems on rubygems to fix the dependency issues.

I have needed to make a change to the modulesync_config Gemfile template as I have needed to use an array of platforms rather than one single platform on the windows gems in order to cover both x86 and x64 mingw platforms along with mswin platform.