(FM-6588) Remove workarounds for puppet < 4.5.2
Since a few days the minitar < 0.6 dependency is being flagged by github's
vulnerability scanner, we remove it now, and drop final pretences at
supporting puppet 3.

Additionally there is some complexity around which specific versions of
the win32 gems to use with puppet 4 < 4.5.2. Since none of those puppet
versions are supported, this summarily drops the workarounds for that
too.