(MODULES-3945) Make travis.yaml more configurable
As part of MODULES-3945 the travis configuration required the use of sudo, dist
and additional before_install steps.  Previously this is was not possible as the
modsync template was not setup.  This commit:
- Allows the sudo: and dist: settings to be overriden from the defaults of
  'false' and not defined respectively.
- Allows modules to specify additional before_install steps.  This is used in
  the powershell module to install Powershell in Travis prior to testing.

Note that this commit is backwards compatible as the new steps are opt-in and
preserve the pre-existing defaults.